requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.23.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
databases==0.4.3
dnspython==2.1.0
email-validator==1.1.3
Faker==8.10.3
fastapi==0.66.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.2
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
orjson==3.6.0
promise==2.3
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
text-unidecode==1.3
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.15.3
watchgod==0.7
websockets==8.1
